requirements.txt
# Core dependencies
mongoengine>=0.24.0
pymongo>=4.0.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
requests>=2.28.0
aiohttp>=3.8.0
pytesseract>=0.3.10
pyyaml>=6.0
selenium>=4.15.0
webdriver-manager>=4.0.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.1
pytest-timeout>=2.1.0
pytest-xdist>=3.3.1
pytest-html>=3.2.0
pytest-benchmark>=4.0.0
pytest-mock>=3.11.1
pytest-clarity>=1.0.1
pytest-sugar>=0.9.7
pytest-instafail>=0.5.0
mongomock>=4.1.2
coverage>=7.2.7
faker>=19.2.0
factory-boy>=3.3.0
freezegun>=1.2.2
responses>=0.23.0
httpretty>=1.1.4
vcrpy>=6.0.0
pytest-vcr>=1.0.2

# Performance testing
locust>=2.17.0
memory-profiler>=0.61.0
py-spy>=0.3.14
pytest-benchmark>=4.0.0

# Code quality
black>=23.7.0
flake8>=6.0.0
pylint>=2.17.4
mypy>=1.4.1
isort>=5.12.0
bandit>=1.7.5
safety>=2.3.5
semgrep>=1.45.0
radon>=6.0.1
xenon>=0.9.0
vulture>=2.9.1
pylint-django>=2.5.5
pylint-flask>=0.6.0

# Security testing
safety>=2.3.5
semgrep>=1.45.0
bandit>=1.7.5
cryptography>=41.0.0

# Pre-commit hooks
pre-commit>=3.3.3
commitizen>=3.13.0

# Monitoring and reporting
psutil>=5.9.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Load testing
locust>=2.17.0
wrk>=0.1.0

# Database testing
testcontainers>=3.7.0
pytest-postgresql>=5.0.0
pytest-mysql>=2.2.0

# API testing
httpx>=0.24.0
pytest-httpx>=0.21.0
requests-mock>=1.11.0

# UI testing
selenium>=4.15.0
webdriver-manager>=4.0.0
playwright>=1.37.0
pytest-playwright>=0.4.0

# Mock and fixtures
pytest-mock>=3.11.1
responses>=0.23.0
httpretty>=1.1.4
vcrpy>=6.0.0
